refactor: extract connector policy analysis to shared utility
