+++ lib/util/helpers.js
@@ -85,3 +85,18 @@ export async function callWithRetry(fn, _description) {
     throw error
   }
 }
+
+/**
+ * Formats a raw string (e.g., SNAKE_CASE status) to Title Case with spaces.
+ * @param {string} s - The raw string
+ * @returns {string} The formatted string
+ */
+export function formatStatus(s) {
+  if (!s) {
+    return 'Unknown'
+  }
+  return String(s)
+    .replace(/_/g, ' ')
+    .toLowerCase()
+    .replace(/\b\w/g, l => l.toUpperCase())
+}
